Refactored SubgroupInfo again, and rewrote construct_objects
Specifically, rewrote construct_objects to read the subgroup membership
information from the View instead of calling the subgroup membership
function again. This means construct_objects doesn't have to catch
subgroup_provisioning_exceptions, just check the View's
is_adequately_provisioned status.

+++ derecho/experiments/local_filewriter_test.cpp
@@ -33,7 +33,7 @@ int main(int argc, char* argv[]) {
     m.size = 1000;
     //    m.message_buffer = std::move(buffer);
 
-    persistence::message message{&buf[0], m.size, 1, m.sender_rank, (uint64_t)m.index};
+    persistence::message message{&buf[0], m.size, 1, (uint32_t)m.sender_rank, (uint64_t)m.index};
 
     file_writer->write_message(message);
